# Conflicts:
#	server/notification-providers/telegram.js
#	src/components/notifications/Telegram.vue
#	src/languages/en.js

+++ server/jobs/clear-old-data.js
@@ -25,15 +25,20 @@ const DEFAULT_KEEP_PERIOD = 180;
         parsedPeriod = DEFAULT_KEEP_PERIOD;
     }
 
-    log(`Clearing Data older than ${parsedPeriod} days...`);
+    if (parsedPeriod < 1) {
+        log(`Data deletion has been disabled as period is less than 1. Period is ${parsedPeriod} days.`);
+    } else {
 
-    try {
-        await R.exec(
-            "DELETE FROM heartbeat WHERE time < DATETIME('now', '-' || ? || ' days') ",
-            [ parsedPeriod ]
-        );
-    } catch (e) {
-        log(`Failed to clear old data: ${e.message}`);
+        log(`Clearing Data older than ${parsedPeriod} days...`);
+
+        try {
+            await R.exec(
+                "DELETE FROM heartbeat WHERE time < DATETIME('now', '-' || ? || ' days') ",
+                [ parsedPeriod ]
+            );
+        } catch (e) {
+            log(`Failed to clear old data: ${e.message}`);
+        }
     }
 
     exit();
